Include annotations when publishing from the cms and when loading into curriculum.
